Rename project for UI tests
