+++ PriceTrackerMobile.sln
@@ -9,7 +9,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PriceTrackerMobile.iOS", "P
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PriceTrackerMobile", "PriceTrackerMobile\PriceTrackerMobile.csproj", "{4F98BD9B-9E85-46C9-B60B-A1A4A2EC832B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrossPlatformTest", "CrossPlatformTest\CrossPlatformTest.csproj", "{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrossPlatformUITest", "CrossPlatformUITest\CrossPlatformUITest.csproj", "{F7CD1DD3-138A-4357-9071-95E350E0C2DC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -75,18 +75,18 @@ Global
 		{4F98BD9B-9E85-46C9-B60B-A1A4A2EC832B}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{4F98BD9B-9E85-46C9-B60B-A1A4A2EC832B}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 		{4F98BD9B-9E85-46C9-B60B-A1A4A2EC832B}.Release|iPhoneSimulator.Deploy.0 = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|iPhone.Build.0 = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{1C074B93-5A37-4AB2-8474-68A3E1E6E2A6}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|iPhone.Build.0 = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{F7CD1DD3-138A-4357-9071-95E350E0C2DC}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
